ECL objects have changed so we need to update the pull request before
merging

+++ framework/pcore/src/RbTuple.cc
@@ -35,6 +35,13 @@ RbTupleManager::~RbTupleManager()
 {
 }
 
+RbTupleManager::RbTupleManager(int nproc, const char* file, const char* workdir)
+{
+  m_filename = file;
+  m_nproc = nproc;
+  m_workdir = workdir;
+}
+
 // Access to Singleton
 RbTupleManager& RbTupleManager::Instance()
 {
@@ -45,15 +52,17 @@ RbTupleManager& RbTupleManager::Instance()
 }
 
 // Global initialization
-void RbTupleManager::init(int nprocess, const char* filename)
+void RbTupleManager::init(int nprocess, const char* filename, const char* workdir)
 {
   m_filename = filename;
   m_nproc = nprocess;
+  m_workdir = workdir;
 
   if (ProcHandler::EvtProcID() == -1 && m_nproc > 0) {
+
     // should be called only from main
     // Open current directory
-    std::string dir = ".";
+    std::string dir = m_workdir;
     DIR* dp;
     struct dirent* dirp;
     if ((dp = opendir(dir.c_str())) == NULL) {
@@ -66,7 +75,8 @@ void RbTupleManager::init(int nprocess, const char* filename)
     while ((dirp = readdir(dp)) != NULL) {
       std::string curfile = std::string(dirp->d_name);
       if (curfile.compare(0, compfile.size(), compfile) == 0) {
-        unlink(curfile.c_str());
+        //        unlink(curfile.c_str());
+        unlink((m_workdir + "/" + curfile).c_str());
       }
     }
     closedir(dp);
@@ -84,7 +94,8 @@ void RbTupleManager::register_module(Module* mod)
 int RbTupleManager::begin(int procid)
 {
   if (m_nproc > 0) {
-    std::string fileNamePlusId = m_filename + '.' + std::to_string(procid);
+    std::string fileNamePlusId = m_workdir + "/" +
+                                 m_filename + '.' + std::to_string(procid);
     m_root = new TFile(fileNamePlusId.c_str(), "update");
     //    printf("RbTupleManager: histo file opened for process %d (pid=%d)\n",
     //           procid, getpid());
@@ -118,8 +129,16 @@ int RbTupleManager::terminate()
   return 0;
 }
 
+int RbTupleManager::dump()
+{
+  if (m_root != NULL) {
+    m_root->Write();
+  }
+  return 0;
+}
+
 // Functions called from main process
-int RbTupleManager::hadd()
+int RbTupleManager::hadd(bool deleteflag)
 {
   // No need to call this function when nprocess=0
   if (m_nproc == 0) {
@@ -127,16 +146,18 @@ int RbTupleManager::hadd()
     return 0;
   }
 
+  printf("RbTupleManager::hadd started\n");
+
   // Set up merger with output file
   TFileMerger merger(false, false);
   if (!merger.OutputFile(m_filename.c_str())) {
-    //    printf ( "RbTupleManager:: error to open output file %s\n", m_filename );
     B2ERROR("RbTupleManager:: error to open output file " << m_filename);
     return -1;
   }
 
   // Open current directory
-  std::string dir = ".";
+  //  std::string dir = ".";
+  std::string dir = m_workdir;
   DIR* dp;
   struct dirent* dirp;
   if ((dp = opendir(dir.c_str())) == NULL) {
@@ -146,13 +167,16 @@ int RbTupleManager::hadd()
 
   std::vector<std::string> filenames;
   // Scan the directory and register all histogram files
-  std::string compfile = std::string(m_filename) + ".";
+  //  std::string compfile = dir + "/" + m_filename + ".";
+  std::string compfile = m_filename + ".";
+  printf("compfile = %s\n", compfile.c_str());
   while ((dirp = readdir(dp)) != NULL) {
     std::string curfile = std::string(dirp->d_name);
+    //    printf("Checking %s with compfile%s\n", curfile.c_str(), compfile.c_str());
     if (curfile.compare(0, compfile.size(), compfile) == 0) {
-      //      printf ( "RbTupleManager:: adding  file =%s\n", curfile.c_str() );
-      merger.AddFile(curfile.c_str());
-      filenames.push_back(curfile);
+      printf("RbTupleManager:: adding  file =%s\n", curfile.c_str());
+      merger.AddFile((m_workdir + "/" + curfile).c_str());
+      filenames.push_back(m_workdir + "/" + curfile);
     }
   }
   closedir(dp);
@@ -164,14 +188,16 @@ int RbTupleManager::hadd()
     return -1;
   }
 
-  // Delete temporary files
-  vector<string>::iterator it;
-  for (it = filenames.begin(); it != filenames.end(); ++it) {
-    string& hfile = *it;
-    unlink(hfile.c_str());
+  if (deleteflag) {
+    // Delete temporary files
+    vector<string>::iterator it;
+    for (it = filenames.begin(); it != filenames.end(); ++it) {
+      string& hfile = *it;
+      unlink(hfile.c_str());
+    }
   }
 
-  //  printf("RbTupleManager: histogram files are added\n");
+  printf("RbTupleManager: histogram files are added\n");
   //  B2INFO ( "RbTupleManager : histogram files are added" );
 
   return 0;
